+++ src/pages/index.js
@@ -1,3 +1,4 @@
+export { PostPage } from "./PostPage"
 export { ProfilePage } from "./ProfilePage"
 export { ExplorePage } from "./ExplorePage"
 export { BookmarkPage } from "./BookmarkPage"
