+++ src/routes/Routes.js
@@ -5,6 +5,7 @@ import {
   ExplorePage,
   HomePage,
   LoginPage,
+  PostPage,
   ProfilePage,
 } from "../pages";
 import { RequiresAuth } from "./requiresAuth";
@@ -20,6 +21,7 @@ export const AllRoutes = () => {
         <Route path="/bookmark" element={<BookmarkPage />} />
         <Route path="/explore" element={<ExplorePage />} />
         <Route path="/profile/:username" element={<ProfilePage />} />
+        <Route path="/post/:postId" element={<PostPage />} />
       </Route>
       {token ? (
         <Route path="/login" element={<Navigate to="/" replace />} />
